refactor: split large guards.ts into domain-specific modules
- Create primitive-guards.ts for basic type validation (250 lines)
- Create api-guards.ts for API and data validation (155 lines)
- Create error-guards.ts for error type validation (72 lines)
- Create dom-guards.ts for DOM and React validation (64 lines)
- Create utility-guards.ts for advanced guard utilities (86 lines)
- Update guards.ts to re-export from focused modules (113 lines)

This addresses the code review feedback about the 593-line guards.ts
file being too large. The new structure organizes guards by domain
(primitive, API, error, DOM, utility) while preserving backward
compatibility through re-exports.
